+++ HolyRepoSkills/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.1.5
+- Update logger plugin name to mod name
+- Update config for custom war cries
+- Update config to disable war cries
+
 ## 1.1.4
 - Readme update